requirements.txt
requests
openai
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
pillow
opencv-python-headless
